split charactor requirements into multiple files
