+++ Content.Shared/Customization/Systems/CharacterRequirements.cs
@@ -1,12 +1,5 @@
-using System.Linq;
-using Content.Shared.CCVar;
-using Content.Shared.Clothing.Loadouts.Prototypes;
-using Content.Shared.Humanoid.Prototypes;
-using Content.Shared.Players.PlayTimeTracking;
 using Content.Shared.Preferences;
 using Content.Shared.Roles;
-using Content.Shared.Roles.Jobs;
-using Content.Shared.Traits;
 using JetBrains.Annotations;
 using Robust.Shared.Configuration;
 using Robust.Shared.Prototypes;
@@ -22,12 +15,15 @@ public abstract partial class CharacterRequirement
 {
     /// <summary>
     ///     If true valid requirements will be treated as invalid and vice versa
+    ///     This inversion is done by other systems like <see cref="CharacterRequirementsSystem"/>, not this one
     /// </summary>
     [DataField]
     public bool Inverted;
 
     /// <summary>
     ///     Checks if this character requirement is valid for the given parameters
+    ///     <br />
+    ///     You should probably not be calling this directly, use <see cref="CharacterRequirementsSystem"/>
     /// </summary>
     /// <param name="reason">Description for the requirement, shown when not null</param>
     public abstract bool IsValid(
@@ -41,440 +37,3 @@ public abstract partial class CharacterRequirement
         out FormattedMessage? reason
     );
 }
-
-
-#region HumanoidCharacterProfile
-
-/// <summary>
-///     Requires the profile to be within an age range
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterAgeRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public int Min;
-
-    [DataField(required: true)]
-    public int Max;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-age-requirement",
-            ("inverted", Inverted), ("min", Min), ("max", Max)));
-        return profile.Age >= Min && profile.Age <= Max;
-    }
-}
-
-/// <summary>
-///   Requires the profile to use either a Backpack, Satchel, or Duffelbag
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterBackpackTypeRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public BackpackPreference Preference;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-backpack-type-requirement",
-            ("inverted", Inverted),
-            ("type", Loc.GetString($"humanoid-profile-editor-preference-{Preference.ToString().ToLower()}"))));
-        return profile.Backpack == Preference;
-    }
-}
-
-/// <summary>
-///     Requires the profile to use either Jumpsuits or Jumpskirts
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterClothingPreferenceRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public ClothingPreference Preference;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-clothing-preference-requirement",
-            ("inverted", Inverted),
-            ("preference", Loc.GetString($"humanoid-profile-editor-preference-{Preference.ToString().ToLower()}"))));
-        return profile.Clothing == Preference;
-    }
-}
-
-/// <summary>
-///     Requires the profile to be a certain species
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterSpeciesRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public List<ProtoId<SpeciesPrototype>> Species;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        const string color = "green";
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-species-requirement",
-            ("inverted", Inverted),
-            ("species", $"[color={color}]{string.Join($"[/color], [color={color}]",
-                Species.Select(s => Loc.GetString(prototypeManager.Index(s).Name)))}[/color]")));
-
-        return Species.Contains(profile.Species);
-    }
-}
-
-/// <summary>
-///     Requires the profile to have one of the specified traits
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterTraitRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public List<ProtoId<TraitPrototype>> Traits;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        const string color = "lightblue";
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-trait-requirement",
-            ("inverted", Inverted),
-            ("traits", $"[color={color}]{string.Join($"[/color], [color={color}]",
-                Traits.Select(t => Loc.GetString($"trait-name-{t}")))}[/color]")));
-
-        return Traits.Any(t => profile.TraitPreferences.Contains(t.ToString()));
-    }
-}
-
-/// <summary>
-///     Requires the profile to have one of the specified loadouts
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterLoadoutRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public List<ProtoId<LoadoutPrototype>> Loadouts;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        const string color = "lightblue";
-        reason = FormattedMessage.FromMarkup(Loc.GetString("character-loadout-requirement",
-            ("inverted", Inverted),
-            ("loadouts", $"[color={color}]{string.Join($"[/color], [color={color}]",
-                Loadouts.Select(l => Loc.GetString($"loadout-name-{l}")))}[/color]")));
-
-        return Loadouts.Any(l => profile.LoadoutPreferences.Contains(l.ToString()));
-    }
-}
-
-#endregion
-
-#region Jobs
-
-/// <summary>
-///     Requires the selected job to be one of the specified jobs
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterJobRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public List<ProtoId<JobPrototype>> Jobs;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        var jobs = new List<FormattedMessage>();
-
-        // Get the job names and department colors
-        foreach (var j in Jobs)
-        {
-            var jobProto = prototypeManager.Index(j);
-            var color = Color.LightBlue;
-
-            foreach (var dept in prototypeManager.EnumeratePrototypes<DepartmentPrototype>()
-                .OrderBy(d => Loc.GetString($"department-{d.ID}")))
-            {
-                if (!dept.Roles.Contains(j))
-                    continue;
-
-                color = dept.Color;
-                break;
-            }
-
-            jobs.Add(FormattedMessage.FromMarkup($"[color={color.ToHex()}]{Loc.GetString(jobProto.Name)}[/color]"));
-        }
-
-        // Join the job names
-        var jobsList = string.Join(", ", jobs.Select(j => j.ToMarkup()));
-        var jobsString = Loc.GetString("character-job-requirement",
-            ("inverted", Inverted), ("jobs", jobsList));
-
-        reason = FormattedMessage.FromMarkup(jobsString);
-        return Jobs.Contains(job.ID);
-    }
-}
-
-/// <summary>
-///     Requires the selected job to be in one of the specified departments
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterDepartmentRequirement : CharacterRequirement
-{
-    [DataField(required: true)]
-    public List<ProtoId<DepartmentPrototype>> Departments;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        var departments = new List<FormattedMessage>();
-
-        // Get the department names and colors
-        foreach (var d in Departments)
-        {
-            var deptProto = prototypeManager.Index(d);
-            var color = deptProto.Color;
-
-            departments.Add(FormattedMessage.FromMarkup($"[color={color.ToHex()}]{Loc.GetString($"department-{deptProto.ID}")}[/color]"));
-        }
-
-        // Join the department names
-        var departmentsList = string.Join(", ", departments.Select(d => d.ToMarkup()));
-        var departmentsString = Loc.GetString("character-department-requirement",
-            ("inverted", Inverted), ("departments", departmentsList));
-
-        reason = FormattedMessage.FromMarkup(departmentsString);
-        return Departments.Any(d => prototypeManager.Index(d).Roles.Contains(job.ID));
-    }
-}
-
-/// <summary>
-///     Requires the playtime for a department to be within a certain range
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterDepartmentTimeRequirement : CharacterRequirement
-{
-    [DataField]
-    public TimeSpan Min = TimeSpan.MinValue;
-
-    [DataField]
-    public TimeSpan Max = TimeSpan.MaxValue;
-
-    [DataField(required: true)]
-    public ProtoId<DepartmentPrototype> Department;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        // Disable the requirement if the role timers are disabled
-        if (!configManager.GetCVar(CCVars.GameRoleTimers))
-        {
-            reason = null;
-            return !Inverted;
-        }
-
-        var department = prototypeManager.Index(Department);
-
-        // Combine all of this department's job playtimes
-        var playtime = TimeSpan.Zero;
-        foreach (var other in department.Roles)
-        {
-            var proto = prototypeManager.Index<JobPrototype>(other).PlayTimeTracker;
-
-            playTimes.TryGetValue(proto, out var otherTime);
-            playtime += otherTime;
-        }
-
-        if (playtime > Max)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-department-too-high",
-                    ("time", playtime.TotalMinutes - Max.TotalMinutes),
-                    ("department", Loc.GetString($"department-{department.ID}")),
-                    ("departmentColor", department.Color)));
-            return false;
-        }
-
-        if (playtime < Min)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-department-insufficient",
-                    ("time", Min.TotalMinutes - playtime.TotalMinutes),
-                    ("department", Loc.GetString($"department-{department.ID}")),
-                    ("departmentColor", department.Color)));
-            return false;
-        }
-
-        reason = null;
-        return true;
-    }
-}
-
-/// <summary>
-///     Requires the player to have a certain amount of overall job time
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterOverallTimeRequirement : CharacterRequirement
-{
-    [DataField]
-    public TimeSpan Min = TimeSpan.MinValue;
-
-    [DataField]
-    public TimeSpan Max = TimeSpan.MaxValue;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        // Disable the requirement if the role timers are disabled
-        if (!configManager.GetCVar(CCVars.GameRoleTimers))
-        {
-            reason = null;
-            return !Inverted;
-        }
-
-        // Get the overall time
-        var overallTime = playTimes.GetValueOrDefault(PlayTimeTrackingShared.TrackerOverall);
-
-        if (overallTime > Max)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-overall-too-high",
-                    ("time", overallTime.TotalMinutes - Max.TotalMinutes)));
-            return false;
-        }
-
-        if (overallTime < Min)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-overall-insufficient",
-                    ("time", Min.TotalMinutes - overallTime.TotalMinutes)));
-            return false;
-        }
-
-        reason = null;
-        return true;
-    }
-}
-
-/// <summary>
-///     Requires the playtime for a tracker to be within a certain range
-/// </summary>
-[UsedImplicitly]
-[Serializable, NetSerializable]
-public sealed partial class CharacterPlaytimeRequirement : CharacterRequirement
-{
-    [DataField]
-    public TimeSpan Min = TimeSpan.MinValue;
-
-    [DataField]
-    public TimeSpan Max = TimeSpan.MaxValue;
-
-    [DataField(required: true)]
-    public ProtoId<PlayTimeTrackerPrototype> Tracker;
-
-    public override bool IsValid(IPrototype prototype, JobPrototype job, HumanoidCharacterProfile profile,
-        Dictionary<string, TimeSpan> playTimes,
-        IEntityManager entityManager, IPrototypeManager prototypeManager, IConfigurationManager configManager,
-        out FormattedMessage? reason)
-    {
-        // Disable the requirement if the role timers are disabled
-        if (!configManager.GetCVar(CCVars.GameRoleTimers))
-        {
-            reason = null;
-            return !Inverted;
-        }
-
-        // Get SharedJobSystem
-        if (!entityManager.EntitySysManager.TryGetEntitySystem(out SharedJobSystem? jobSystem))
-        {
-            DebugTools.Assert("CharacterRequirements: SharedJobSystem not found");
-            reason = null;
-            return false;
-        }
-
-        // Get the JobPrototype of the Tracker
-        var trackerJob = jobSystem.GetJobPrototype(Tracker);
-        var jobStr = prototypeManager.Index<JobPrototype>(trackerJob).LocalizedName;
-
-        // Get the primary department of the Tracker
-        if (!jobSystem.TryGetPrimaryDepartment(trackerJob, out var department) &&
-            !jobSystem.TryGetDepartment(trackerJob, out department))
-        {
-            DebugTools.Assert($"CharacterRequirements: Department not found for job {trackerJob}");
-            reason = null;
-            return false;
-        }
-
-        // Get the time for the tracker
-        var time = playTimes.GetValueOrDefault(Tracker);
-        reason = null;
-
-        if (time > Max)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-role-too-high",
-                    ("time", time.TotalMinutes - Max.TotalMinutes),
-                    ("job", jobStr),
-                    ("departmentColor", department.Color)));
-            return false;
-        }
-
-        if (time < Min)
-        {
-            // Show the reason if invalid
-            reason = Inverted
-                ? null
-                : FormattedMessage.FromMarkup(Loc.GetString("character-timer-role-insufficient",
-                    ("time", Min.TotalMinutes - time.TotalMinutes),
-                    ("job", jobStr),
-                    ("departmentColor", department.Color)));
-            return false;
-        }
-
-        return true;
-    }
-}
-
-#endregion
